refactor: rename image-crud to imageId-crud and update related imports
